Refactore GLFW Patch
